[ IntrensicResize ] v1.2 : Refactor 'hdr.corehdr' library and 'hdr.camera' library --------------------------------------------------------------------------------------- TODO : Some implementation modification of hdr.camera lib.
