+++ app/src/main/res/layout/activity_camera.xml
@@ -14,9 +14,7 @@
  limitations under the License.
 -->
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#000"
-    tools:context="com.adrino.renderscript.CameraActivity" />
+    android:background="#000"/>
