+++ hdr/src/main/java/com/adrino/hdr/camera/Constants.java
@@ -1,0 +1,5 @@
+package com.adrino.hdr.camera;
+
+public class Constants {
+    public static int[] EXPOSURE_BRACKET = {10, 0, -10};
+}
