+++ hdr/src/main/res/drawable/ic_navigate_next_black_24dp.xml
