+++ hdr/src/main/res/drawable/ic_photo_camera_black_24dp.xml
