+++ hdr/src/main/res/layout-land/fragment_camera2_basic.xml
@@ -18,7 +18,7 @@
     android:layout_height="match_parent"
     android:orientation="horizontal">
 
-    <com.adrino.renderscript.AutoFitTextureView
+    <com.adrino.hdr.camera.AutoFitTextureView
         android:id="@+id/texture"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
