+++ hdr/src/main/res/layout/fragment_camera2_basic.xml
@@ -17,7 +17,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.adrino.renderscript.AutoFitTextureView
+    <com.adrino.hdr.camera.AutoFitTextureView
         android:id="@+id/texture"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
