+++ hdr/src/main/res/values/colors.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="colorPrimary">#2196f3</color>
+    <color name="colorPrimaryDark">#1769aa</color>
+    <color name="colorAccent">#4dabf5</color>
+    <color name="white">#ffffff</color>
+    <color name="control_background">#cc4285f4</color>
+</resources>
